Change module API responses.
Each endpoint now returns a standardized response in JSON format with the fields "data" (basic data) and "meta" (metadata for this response). If "data" represents a list and the query parameters "page" and "per_page" were passed, then the "pagination" field is added.

To work with response, the following were created:
- src/backend/utils/response_func.py (functions for building a response)
- src/backend/utils/schemas.py (diagrams for building a response)

For convenient work with project launch teams, the following has been added:
- Makefile