Add the include Query parameter to the endpoints Groups, Models, Tasks
Optional include Query parameters have been added to most of the endpoints of our API, making it possible to request more detailed information for the Client.  This implementation is fully compatible with the pagination and JsonResponse construction of our API.
In the current implementation, you can request:

1)Groups:
- accounts
- modules (everything below is only enabled when modules is enabled)
- tasks (everything below is only enabled when tasks is enabled)
- user_replies

2)Modules:
- tasks (everything below is only enabled when tasks is enabled)
- user_replies

3)Tasks
- user_replies

These fields are passed as a list of parameter strings selected from the source string in the format:
http://0.0.0.0:8000/api/groups/?include=modules&include=tasks&include=accounts&include=user_replies

In addition, the JsonResponse structure has been rewritten. Implemented a builder that reduces code repetition