Add an updated task to celery for verifying solutions with code
The functionality for working with the API of entities such as:
- tests
- tasks with a code
- solutions with a code

A background task has been added to celery that performs user code validation in isolated docker containers. The code passes through the tests, with a comparison of the output data.

A status field has been added to the custom solutions model with the code. It is indicated depending on the result of passing the task tests.